adding controller entity service and repo for relationship and property

+++ src/main/java/com/csye7250/project/webapp/repository/NodeRepository.java
@@ -1,9 +1,10 @@
 package com.csye7250.project.webapp.repository;
 
-import com.csye7250.project.webapp.entity.Node;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NodeRepository extends CrudRepository<Node, Long> {
+import com.csye7250.project.webapp.entity.Node;
+
+public interface NodeRepository extends CrudRepository<Node, Integer> {
 
 
 }
